rename : two file name change (handler.js, selected-position-change.js)
from selected-position-change.js to text-selected-position-change.js
from handler.js to weekly-webtoon-handler.js
