+++ text-selected-position-change.js
